+++ apps/common/src/main/scala/com/daml/network/environment/PackageIdResolver.scala
@@ -241,4 +241,11 @@ object PackageIdResolver {
 
   def supportsValidatorLicenseActivity(now: CantonTimestamp, amuletRules: AmuletRules): Boolean =
     supportsValidatorLicenseMetadata(now, amuletRules)
+
+  def supportsPruneAmuletConfigSchedule(now: CantonTimestamp, amuletRules: AmuletRules): Boolean = {
+    val currentConfig = AmuletConfigSchedule(amuletRules).getConfigAsOf(now)
+    val dsoGovernanceVersion =
+      PackageVersion.assertFromString(currentConfig.packageConfig.dsoGovernance)
+    dsoGovernanceVersion >= DarResources.dsoGovernance_0_1_5.metadata.version
+  }
 }