+++ apps/validator/src/main/scala/com/daml/network/validator/automation/ValidatorLicenseActivityTrigger.scala
@@ -8,7 +8,6 @@ import com.daml.network.util.AssignedContract
 import com.daml.network.validator.store.ValidatorStore
 import com.digitalasset.canton.data.CantonTimestamp
 import com.digitalasset.canton.logging.pretty.{Pretty, PrettyPrinting}
-// import com.digitalasset.canton.time.Clock
 import com.digitalasset.canton.tracing.TraceContext
 import io.opentelemetry.api.trace.Tracer
 import org.apache.pekko.stream.Materializer